+++ sickrage/providers/torrent/torrentleech.py
@@ -18,6 +18,7 @@
 
 from __future__ import unicode_literals
 
+import math
 from urlparse import urljoin
 
 from requests.utils import dict_from_cookiejar
@@ -105,7 +106,22 @@ class TorrentLeechProvider(TorrentProvider):
 
                 try:
                     data = self.session.get(search_url).json()
-                    results += self.parse(data, mode)
+
+                    # Handle pagination
+                    num_found = data.get('numFound', 0)
+                    per_page = data.get('perPage', 35)
+
+                    try:
+                        pages = int(math.ceil(100 / per_page))
+                    except ZeroDivisionError:
+                        pages = 1
+
+                    for page in range(2, pages + 1):
+                        results += self.parse(data, mode)
+
+                        if num_found and num_found > per_page and pages > 1:
+                            page_url = urljoin(search_url, 'page/{}/'.format(page))
+                            data = self.session.get(page_url).json()
                 except Exception:
                     sickrage.app.log.debug("No data returned from provider")